fix comparsion errors

+++ PolyEngine/Tests/CoreTests/Src/PriorityQueueTests.cpp
@@ -49,7 +49,7 @@ TEST_CASE("PriorityQueue random push test", "[PriorityQueue]")
 		CHECK(q.size() == i + 1);
 	}
 
-	for (int i = 0; i < testSize - 2; ++i)
+	for (size_t i = 0; i < testSize - 2; ++i)
 	{
 		size_t v1 = q.top();
 		q.pop();
@@ -73,7 +73,7 @@ TEST_CASE("PriorityQueue custom comparator test", "[PriorityQueue]")
 		CHECK(q.size() == i + 1);
 	}
 
-	for (int i = 0; i < testSize - 2; ++i)
+	for (size_t i = 0; i < testSize - 2; ++i)
 	{
 		size_t v1 = q.top();
 		q.pop();
@@ -91,7 +91,7 @@ TEST_CASE("PriorityQueue heap sort", "[PriorityQueue]")
 
 	::pe::core::storage::PriorityQueue<int> q(std::move(table));
 
-	for (int i = 0; i < testSize - 2; ++i)
+	for (size_t i = 0; i < testSize - 2; ++i)
 	{
 		size_t v1 = q.top();
 		q.pop();
